+++ src/components/formularios/index.ts
@@ -1,3 +1,4 @@
 export { default as AfectadosForm } from "./AfectadosForm.vue";
 export { default as VinculadosForm } from "./VinculadosForm.vue";
 export { default as EfectosForm } from "./EfectosForm.vue";
+export { default as PersonalIntervinienteComponent } from "./PersonalIntervinienteComponent.vue";
